feat(catalog-service):Add package entity and DTO
