--- catalog_service/src/main/java/com/escuelajavag4/catalogservice/dto/CategoryDto.java
@@ -1,4 +1,0 @@
-package com.escuelajavag4.catalogservice.dto;
-
-public class CategoryDto {
-}
