--- catalog_service/src/main/java/com/escuelajavag4/catalogservice/model/Category.java
@@ -1,4 +1,0 @@
-package com.escuelajavag4.catalogservice.model;
-
-public class Category {
-}
