+++ catalog_service/src/main/java/com/escuelajavag4/catalogservice/model/dto/MarcaDto.java
@@ -1,0 +1,4 @@
+package com.escuelajavag4.catalogservice.model.dto;
+
+public class MarcaDto {
+}
