+++ catalog_service/src/main/java/com/escuelajavag4/catalogservice/model/dto/ProductDto.java
@@ -1,0 +1,4 @@
+package com.escuelajavag4.catalogservice.model.dto;
+
+public class ProductDto {
+}
