+++ catalog_service/src/main/java/com/escuelajavag4/catalogservice/model/entity/Marca.java
@@ -1,0 +1,4 @@
+package com.escuelajavag4.catalogservice.model.entity;
+
+public class Marca {
+}
